requirements.txt
# Core Flask dependencies
Flask==2.3.2
Flask-Cors==5.0.0
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.12
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
flask-talisman==1.1.0
Werkzeug==2.3.6

# Database
SQLAlchemy==2.0.38
psycopg2==2.9.10
alembic==1.15.2

# Serialization
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.1

# AWS and S3
boto3==1.37.37
botocore==1.37.37
watchtower==3.4.0

# Image processing
Pillow==11.3.0

# Environment and configuration
python-dotenv==1.0.1
click==8.1.8

# AI and Supabase (compatible versions)
google-generativeai==0.8.5
supabase==2.17.0

# Security and authentication
PyJWT==2.10.1
cryptography==44.0.2

# Utilities
requests==2.32.3
PyYAML==6.0.1
Jinja2==3.1.4
MarkupSafe==3.0.2
itsdangerous==2.2.0
blinker==1.9.0
packaging==24.2
setuptools==75.6.0

# Data export
pandas
openpyxl==3.1.2

# Production server
gunicorn==21.2.0